refactor: move utils functions to separate folders
