+++ js/render-photos.js
@@ -1,4 +1,4 @@
-import { createPictureComponent, renderList } from './utils.js';
+import { createPictureComponent, renderList } from './utils';
 
 const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
