+++ js/utils/DOM-manipulation.js
@@ -1,35 +1,3 @@
-import { KeyEnum, STOP_ESCAPING_MODAL_ELEMENTS } from './constants';
-
-export const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-};
-
-export const createRandomIdFromRangeGenerator = (min, max) => {
-  const previousValues = [];
-
-  return () => {
-    let currentValue = getRandomInteger(min, max);
-
-    if (previousValues.length >= (max - min + 1)) {
-      // eslint-disable-next-line no-console
-      console.error(`Перебраны все числа из диапазона от ${min} до ${max}`);
-      return null;
-    }
-
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-
-    previousValues.push(currentValue);
-
-    return currentValue;
-  };
-};
-
 export const createPictureComponent = (photo, template) => {
   const { id = 1, url = '', description = '', likes = 0, comments = [] } = photo;
 
@@ -65,8 +33,6 @@ export const createCommentComponent = (comment) => {
   return listElement;
 };
 
-export const getPhotoById = (id, photos) => photos.find((photo) => photo.id === +id);
-
 export const renderList = (list, containerElement) => (createComponent, template = '') => {
   const fragment = document.createDocumentFragment();
 
@@ -75,12 +41,6 @@ export const renderList = (list, containerElement) => (createComponent, template
   containerElement.append(fragment);
 };
 
-export const isEscapeKey = (evt) => evt.key === KeyEnum.ESCAPE;
-
 export const lockBodyScroll = () => document.body.classList.add('modal-open');
 
 export const unlockBodyScroll = () => document.body.classList.remove('modal-open');
-
-export const isTargetInputField = (target) => STOP_ESCAPING_MODAL_ELEMENTS.some((element) => target.classList.contains(element));
-
-export const hasUniqueValues = (array) => array.length === new Set(array).size;
