+++ js/utils/index.js
@@ -1,0 +1,3 @@
+export * from './checks.js';
+export * from './common.js';
+export * from './DOM-manipulation.js';
